Build system fixes
*Compile schemas in the right directory

Compile the GLib schemas in the installation directory rather than in
an arbitrary directory that just happen to be the common system one.
This allows to install outside a system directory seamlessly.

We also avoid compiling the schemas with a non-empty DESTDIR, because
that means we can't actually update anything useful as the target path
is not the installation one; so in this case we leave this step to the
user -- which in this case is usually a distribution package, which
knows how to deal with that anyway.

Also, drop the now useless `--enable-schemas_compile` option that was
added to workaround distcheck failures.

* Cleanup how data files are distributed

* meson: Generate MATE.desktop

* Don't track or distribute generated .desktop

* Use git.mk to autogenerate .gitignore and enable make maintainer-clean